pyproject.toml
[project]
name = "graphene-blog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=26.3.1",
    "django>=6.0.4",
    "flake8>=7.3.0",
    "graphene-django>=3.2.3",
    "isort>=8.0.1",
    "pre-commit>=4.6.0",
    "python-decouple>=3.8",
]
